Do not cache bindeps config, do runtime discovery
