--- deps/.gitignore
@@ -1,6 +1,0 @@
-build.log
-runtime/
-usr/
-ext.jl
-ext.jl.bak
-deps.jl
